add_scale {ggvis}	R Documentation

Add arbitrary scales to ggvis.

Description

Add arbitrary scales to ggvis.

Usage

add_scale(vis, scale, data_domain = TRUE)

Arguments

vis
Visualisation to modify.

scale
Scale object

data_domain
Should the domain be controlled by a data set which is added to the spec? Should only be set to FALSE in special cases.

domain
Either a vector with static values for the domain, or a reactive that returns a such a vector.

[Package ggvis version 0.4.5 Index]