as.vega {ggvis}	R Documentation

Coerce an ggvis object to a vega list.

Description

This generic function powers the coercion of ggvis objects to vega compatible data structures.

Usage

as.vega(x, ...)

## S3 method for class 'ggvis'
as.vega(x, session = NULL, dynamic = FALSE, ...)

Arguments

x
an object to convert to vega

session
a session object from shiny

dynamic
whether to generate dynamic or static spec

Value

a list. When converted to JSON, will be the type of structure that vega expects.

[Package ggvis version 0.4.5 Index]