auto_group {ggvis}	R Documentation

Automatically group data by grouping variables

Description

Use auto_group to group up a dataset on all categorical variables specified by props, and have each piece rendered by the same mark.

Usage

auto_group(vis, exclude = NULL)

Arguments

vis
The ggvis visualisation to modify.

exclude
A vector containing names of props to exclude from auto grouping. It is often useful to exclude c("x", "y"), when one of those variables is categorical.

See Also

To manually specify grouping variables, see group_by.

Examples

# One line
mtcars %>% ggvis(~disp, ~mpg, stroke = ~factor(cyl)) %>% layer_paths()
# One line for each level of cyl
mtcars %>% ggvis(~disp, ~mpg, stroke = ~factor(cyl)) %>% group_by(cyl) %>%
  layer_paths()
mtcars %>% ggvis(~disp, ~mpg, stroke = ~factor(cyl)) %>% auto_group() %>%
  layer_paths()

# The grouping column can already be stored as a factor
mtcars2 <- mtcars
mtcars2$cyl <- factor(mtcars2$cyl)
mtcars2 %>% ggvis(~disp, ~mpg, stroke = ~cyl) %>% auto_group() %>%
  layer_paths()

[Package ggvis version 0.4.5 Index]