axis_props {ggvis}	R Documentation

Create an axis_props object for controlling axis properties.

Description

The items in each of the lists can be a literal value, like 5 or "blue", or they can be a scaled_value object.

Usage

axis_props(ticks = NULL, majorTicks = NULL, minorTicks = NULL,
  grid = NULL, labels = NULL, title = NULL, axis = NULL)

Arguments

ticks
A named list of line properties for ticks.

majorTicks
A named list of line properties for major ticks.

minorTicks
A named list of line properties for minor ticks.

grid
A named list of line properties for grid lines.

labels
A named list of text properties for axis labels.

title
A named list of text properties for the axis title.

axis
A named list of line properties for the axis line.

[Package ggvis version 0.4.5 Index]