bin_vector {ggvis}	R Documentation

Bin vectors

Description

A generic and several implementations for binning vectors.

Usage

bin_vector(x, weight = NULL, ...)

Arguments

x
A vector to bin

weight
If specified, an integer vector of the same length as x representing the number of occurances of each value in x

...
additional arguments passed through to methods.

width
The width of a bin

origin
The left-most value for bins.

closed
One of "right" or "left" indicating whether right or left edges of bins are included in the bin.

pad
A logical indicating whether the bins should be padded to include an empty bin on each side.

[Package ggvis version 0.4.5 Index]