cocaine {ggvis}	R Documentation

Cocaine seizures in the US.

Description

This dataset comes from STRIDE, the System to Retrieve Information from Drug Evidence. It contains all concaine seizures in the US from 2007 that have a known weight.

Usage

cocaine

Format

Data frame with 3380 observations of 5 variables.

Variables

state

State where seizure occured.

potency

Purity of cocaine, as percentage (100% = pure cocaine, 0% = all filler)

weight

Weight, in grams.

month

Month in which seizure occured.

price

Estimated value in USD.

Use

Use of this data requires your agreement to refer to your analyses as "unvalidated DEA data and to claim authorship and responsibility for any inferences and/or conclusions you may draw from this information."

[Package ggvis version 0.4.5 Index]