ggvisControlOutput {ggvis}	R Documentation

Create a ggvis control output element in UI

Description

This is effectively the same as uiOutput, except that on the client side it may call some plot resizing functions after new controls are drawn.

Usage

ggvisControlOutput(outputId, plotId = NULL)

Arguments

outputId
The output variable to read the value from.

plotId
An optional plot ID or vector of plot IDs. The plots will have their .onControlOutput functions called after the controls are drawn.

Details

ggvisControlOutput is intended to be used with bind_shiny on the server side.

Examples

ggvisControlOutput("plot1")

[Package ggvis version 0.4.5 Index]