ggvisOutputElements {ggvis}	R Documentation

Create HTML elements for ggvis output

Description

This is an internal-facing function similar to ggvisOutput, but with more options.

Usage

ggvisOutputElements(plot_id = rand_id("plot_id"), spec = NULL,
  shiny = TRUE)

Arguments

plot_id
Unique identifier to use for the div containing the ggvis plot.

spec
Plot specification, used internally.

shiny
Should this include headers for Shiny? For dynamic and interactive plots, this should be TRUE; otherwise FALSE.

[Package ggvis version 0.4.5 Index]