handle_brush {ggvis}	R Documentation

Handle brush events on a visualisation.

Description

Currently for brush events to be triggered on a visualisation, you must use a .brush property. This limitation will be lifted in the future.

Usage

handle_brush(vis, on_move = NULL, fill = "black")

Arguments

vis
Visualisation to listen to.

on_move
Callback function with arguments:

items

A data frame containing information about the items under the plot. An empty data.frame if no points under the brush.

page_loc

Location of the brush with repsect to the page

plot_loc

Location of the brush with respect to the plot

session

The session, used to communicate with the browser

fill
Colour of the brush.

Examples

# Display tooltip when objects are brushed
mtcars %>%
  ggvis(x = ~wt, y = ~mpg, size.brush := 400) %>%
  layer_points() %>%
  handle_brush(function(items, page_loc, session, ...) {
    show_tooltip(session, page_loc$r + 5, page_loc$t, html = nrow(items))
  })

[Package ggvis version 0.4.5 Index]