input_slider {ggvis}	R Documentation

Create an interactive slider.

Description

Create an interactive slider.

Usage

input_slider(min, max, value = (min + max)/2, step = NULL,
  round = FALSE, format = NULL, locale = "us", ticks = TRUE,
  animate = FALSE, sep = ",", pre = NULL, post = NULL,
  label = "", id = rand_id("slider_"), map = identity)

Arguments

min
The minimum value (inclusive) that can be selected.

max
The maximum value (inclusive) that can be selected.

value
The initial value of the slider. A numeric vector of length one will create a regular slider; a numeric vector of length two will create a double-ended range slider. A warning will be issued if the value doesn't fit between min and max.

step
Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size). If the values are dates, step is in days; if the values are times (POSIXt), step is in seconds.

round
TRUE to round all values to the nearest integer; FALSE if no rounding is desired; or an integer to round to that number of digits (for example, 1 will round to the nearest 10, and -2 will round to the nearest .01). Any rounding will be applied after snapping to the nearest step.

format
Deprecated.

locale
Deprecated.

ticks
FALSE to hide tick marks, TRUE to show them according to some simple heuristics.

animate
TRUE to show simple animation controls with default settings; FALSE not to; or a custom settings list, such as those created using animationOptions.

sep
Separator between thousands places in numbers.

pre
A prefix string to put in front of the value.

post
A suffix string to put after the value.

label
Display label for the control, or NULL for no label.

id
A unique identifier for this input. Usually generated automatically.

map
A function with single argument x, the value of the control on the client. Returns a modified value.

See Also

Other interactive input: input_checkbox, input_select, input_text

Examples

input_slider(0, 100)
input_slider(0, 100, label = "binwidth")
input_slider(0, 100, value = 50)

# Supply two values to value to make a double-ended sliders
input_slider(0, 100, c(25, 75))

# You can use map to transform the outputs
input_slider(-5, 5, label = "Log scale", map = function(x) 10 ^ x)

[Package ggvis version 0.4.5 Index]