layer_densities {ggvis}	R Documentation

Transformation: density estimate

Description

transform_density is a data transformation that computes a kernel density estimate from a dataset. layer_density combines transform_density with mark_path and mark_area to display a smooth line and its standard errror.

Usage

layer_densities(vis, ..., kernel = "gaussian", adjust = 1,
  density_args = list(), area = TRUE)

Arguments

vis
The visualisation to modify

...
Visual properties, passed on to props.

kernel
Smoothing kernel. See density for details.

adjust
Multiple the default bandwidth by this amount. Useful for controlling wiggliness of density.

density_args
Other arguments passed on to compute_density and thence to density.

area
Should there be a shaded region drawn under the curve?

Examples

# Basic density estimate
faithful %>% ggvis(~waiting) %>% layer_densities()
faithful %>% ggvis(~waiting) %>% layer_densities(area = FALSE)

# Control bandwidth with adjust
faithful %>% ggvis(~waiting) %>% layer_densities(adjust = .25)
faithful %>% ggvis(~waiting) %>%
  layer_densities(adjust = input_slider(0.1, 5))

# Control stroke and fill
faithful %>% ggvis(~waiting) %>%
  layer_densities(stroke := "red", fill := "red")

# With groups
PlantGrowth %>% ggvis(~weight, fill = ~group) %>% group_by(group) %>%
  layer_densities()
PlantGrowth %>% ggvis(~weight, stroke = ~group) %>% group_by(group) %>%
  layer_densities(strokeWidth := 3, area = FALSE)

[Package ggvis version 0.4.5 Index]