layer_f {ggvis}	R Documentation

Create a new layering function.

Description

The layer function is run, and then the state before the code was run is restored - this allows layers to be effectively isolated from the rest of the plot.

Usage

layer_f(vis, fun)

Arguments

vis
The ggvis visualisation to modify.

fun
A function that takes a single argument, the current visualisation as input, and returns a modified visualisation.

Examples

mtcars %>% ggvis(~mpg) %>%
  layer_f(function(v) {
     v %>% compute_bin(~mpg) %>% layer_points(x = ~x_, y = ~count_)
  }) %>%
  layer_points(y = ~wt)

[Package ggvis version 0.4.5 Index]