layer_lines {ggvis}	R Documentation

Layer lines on a plot.

Description

layer_lines differs from layer_paths in that layer_lines sorts the data on the x variable, so the line will always proceed from left to right, whereas layer_paths will draw a line in whatever order appears in the data.

Usage

layer_lines(vis, ...)

Arguments

vis
Visualisation to modify.

...
Visual properties.

See Also

layer_paths

Examples

mtcars %>% ggvis(~wt, ~mpg, stroke = ~factor(cyl)) %>% layer_lines()

# Equivalent to
mtcars %>% ggvis(~wt, ~mpg, stroke = ~factor(cyl)) %>%
  group_by(cyl) %>% dplyr::arrange(wt) %>% layer_paths()

[Package ggvis version 0.4.5 Index]