legend_props {ggvis}	R Documentation

Create an axis_props object for controlling legend properties.

Description

The items in each of the lists can be a literal value, like 5 or "blue", or they can be a scaled_value object.

Usage

legend_props(title = NULL, labels = NULL, symbols = NULL,
  gradient = NULL, legend = NULL)

Arguments

title
A named list of text properties for the legend title.

labels
A named list of text properties for legend labels.

symbols
A named list of line properties for symbols (for discrete legend items).

gradient
A named list of line properties a continuous color gradient.

legend
A named list of line properties for the overall legend. The x and y position can be set here, which will override automatic positioning.

[Package ggvis version 0.4.5 Index]