linked_brush {ggvis}	R Documentation

Create a linked brush object.

Description

A linked brush has two sides: input and output

Usage

linked_brush(keys, fill = "red")

Arguments

keys
vector of all possible keys, if known.

fill
brush colour

Value

A list with components:

input
A function that takes a visualisation as an argument and adds an input brush to that plot

selected
A reactive providing a logical vector that describes which points are under the brush

fill
A reactive that gives the fill colour of points under the brush

Note

linked_brush is very new and is likely to change substantially in the future

Examples

lb <- linked_brush(keys = 1:nrow(mtcars), "red")

# Change the colour of the points
mtcars %>%
 ggvis(~disp, ~mpg) %>%
 layer_points(fill := lb$fill, size.brush := 400) %>%
 lb$input()

# Display one layer with all points and another layer with selected points
library(shiny)
mtcars %>%
 ggvis(~disp, ~mpg) %>%
 layer_points(size.brush := 400) %>%
 lb$input() %>%
 layer_points(fill := "red", data = reactive(mtcars[lb$selected(), ]))

[Package ggvis version 0.4.5 Index]