mark {ggvis}	R Documentation

Create a new "mark" object.

Description

A mark object is a close mapping to a vega mark object. Vega marks are documented in https://github.com/trifacta/vega/wiki/Marks.

Usage

mark(type, props, data)

is.mark(x)

Arguments

type
A string with the vega type.

props
A list of properties, created by props.

data
A reactive data object.

Details

This function is designed to be used by authors of new types of mark.

[Package ggvis version 0.4.5 Index]