new_prop {ggvis}	R Documentation

Create new prop object

Description

The resulting object has the following fields:

Usage

new_prop(x, property, scale, offset, mult, env, event, label)

Details

property The name of a visual property, like "x", "x2", "width", "y", "fill".

value A value. Can be a constant, reactive, or quoted expression.

scale A string with name of a scale. Typically something like "x", "y", "fill", but can also be a custom name like "foo".

offset Additive pixel offset used to adjust scaled values.

mult Multiplicative pixel offset used to adjust scaled values.

event A event like "update", "enter", "exit", "hover", "brush".

env An environment in which to evaluate a variable or reactive value.

[Package ggvis version 0.4.5 Index]