%>% {ggvis}	R Documentation

Pipe graphics

Description

Like dplyr, ggvis also uses the pipe function, %>% to turn function composition into a series of imperative statements.

Arguments

lhs, rhs
A visualisation and a function to apply to it

Examples

# Instead of
layer_points(ggvis(mtcars, ~mpg, ~wt))
# you can write
mtcars %>% ggvis(~mpg, ~wt) %>% layer_points()

[Package ggvis version 0.4.5 Index]