propname_to_scale {ggvis}	R Documentation

Convert the name of a property to the name of its default scale.

Description

This is mainly used to ensure that similar properties share the same scale by default - e.g. x and x2 should use the same scale.

Usage

propname_to_scale(prop)

Arguments

prop
character vector of property names. Any unrecognised names are left unchanged.

Value

character vector of default scale names.

Examples

propname_to_scale(c("x", "x2"))
propname_to_scale(c("foo", "bar"))
propname_to_scale(c("opacity", "fillOpacity", "strokeOpacity"))

[Package ggvis version 0.4.5 Index]