scaletype_to_vega_scaletype {ggvis}	R Documentation

Given the type of a ggvis scale, get the name of its corresponding vega scale

Description

Given the type of a ggvis scale, get the name of its corresponding vega scale

Usage

scaletype_to_vega_scaletype(type)

Arguments

type
property type: numeric, ordinal, nominal, logical or datetime.

[Package ggvis version 0.4.5 Index]