set_options {ggvis}	R Documentation

Set options for a ggvis plot

Description

Set options for a ggvis plot

Usage

set_options(vis, width = NULL, height = NULL, keep_aspect = NULL,
  resizable = NULL, padding = NULL, duration = NULL,
  renderer = NULL, hover_duration = NULL)

Arguments

vis
Visualisation to modify

width, height
Width and height of plot, in pixels. Default is 600x400. width or height can also be "auto", in which case the plot will size to fit in the containing div. This is useful only in a Shiny app or custom HTML output. Note that height="auto" should only be used when the plot is placed within a div that has a fixed height; if not, automatic height will not work, due to the way that web browsers do vertical layout.

keep_aspect
Should the aspect ratio be preserved? The default value is FALSE, or the value of getOption("ggvis.keep_aspect"), if it is set.

resizable
If TRUE, allow the user to resize the plot. The default value is TRUE, or the value of getOption("ggvis.resizable"), if it is set. Not compatible when width or height is "auto".

padding
A padding object specifying padding on the top, right, left, and bottom. See padding.

duration
Duration of transitions, in milliseconds.

renderer
The renderer to use in the browser. Can be "canvas" or "svg" (the default).

hover_duration
The amount of time for hover transitions, in milliseconds.

See Also

getOption and options, for getting and setting global options.

default_options to see the default options.

Examples

mtcars %>%
  ggvis(~wt, ~mpg) %>%
  layer_points() %>%
  set_options(width = 300, height = 200, padding = padding(10, 10, 10, 10))

# Display the default options
str(default_options())

[Package ggvis version 0.4.5 Index]