show_tooltip {ggvis}	R Documentation

Send a message to the client to show or hide a tooltip

Description

Send a message to the client to show or hide a tooltip

Usage

show_tooltip(session, l = 0, t = 0, html = "")

hide_tooltip(session)

Arguments

session
A Shiny session object.

l
Pixel location of left edge of tooltip (relative to page)

t
Pixel location of top edge of tooltip (relative to page)

html
HTML to display in the tooltip box.

[Package ggvis version 0.4.5 Index]