sidebarBottomPage {ggvis}	R Documentation

Create a page with a sidebar

Description

This creates a page with a sidebar, where the sidebar moves to the bottom when the width goes below a particular value.

Usage

sidebarBottomPage(sidebarPanel, mainPanel, shiny_headers = TRUE)

sidebarBottomPanel(...)

mainTopPanel(...)

Arguments

sidebarPanel
The sidebarBottomPanel containing input controls.

mainPanel
The mainTopPanel containing the main content.

shiny_headers
Should Shiny headers be embedded in the page? This should be TRUE for interactive/dynamic pages, FALSE for static pages.

...
Additional tags.

Examples

sidebarBottomPage(sidebarBottomPanel(), mainTopPanel())

[Package ggvis version 0.4.5 Index]