singular {ggvis}	R Documentation

singular.

Description

Use singular when you want constant x or y position.

Usage

singular()

scale_singular(vis, property, name = property, label = name,
  points = TRUE, domain = NULL, override = NULL)

Arguments

vis
A ggvis object.

property
The name of a property, such as "x", "y", "fill", "stroke", etc.

name
Name of the scale, such as "x", "y", "fill", etc. Can also be an arbitrary name like "foo".

label
Label for the scale. Used for axis or legend titles.

points
If TRUE (default), distributes the ordinal values over a quantitative range at uniformly spaced points. The spacing of the points can be adjusted using the padding property. If FALSE, the ordinal scale will construct evenly-spaced bands, rather than points. Note that if any mark is added with a band() prop, then the scale for that prop will automatically have points set to FALSE.

domain
The domain of the scale, representing the set of data values. For ordinal scales, a character vector; for quantitative scales, a numeric vector of length two. Either value (but not both) may be NA, in which case domainMin or domainMax is set. For dynamic scales, this can also be a reactive which returns the appropriate type of vector.

override
Should the domain specified by this ggvis_scale object override other ggvis_scale objects for the same scale? Useful when domain is manually specified. For example, by default, the domain of the scale will contain the range of the data, but when this is TRUE, the specified domain will override, and the domain can be smaller than the range of the data. If FALSE, the domain will not behave this way. If left NULL, then it will be treated as TRUE whenever domain is non-NULL.

Examples

mtcars %>% ggvis("", ~mpg) %>%
  layer_points() %>%
  scale_nominal("x") %>%
  add_axis("x", title = "", tick_size_major = 0)

# OR
mtcars %>% ggvis("", ~mpg) %>%
  layer_points() %>%
  scale_singular("x")

# OR, even simpler
mtcars %>% ggvis(singular(), ~mpg) %>% layer_points()

# In the other direction:
mtcars %>% ggvis(~mpg, singular()) %>% layer_points()

[Package ggvis version 0.4.5 Index]