vector_type {ggvis}	R Documentation

Determine the "type" of a vector

Description

The vector_type collapses down the class of base vectors into something useful more for visualisation, yielding one of "datetime", "numeric", "ordinal", "nominal" or "logical".

Usage

vector_type(x)

Arguments

x
a vector

See Also

default_scale, which uses this when picking the default scale.

[Package ggvis version 0.4.5 Index]