vega_data_parser {ggvis}	R Documentation

Determine the vega data type for a vector

Description

This is used to specify the data type so that the appropriate parser is used when Vega receives the data.

Usage

vega_data_parser(x)

Arguments

x
A vector.

[Package ggvis version 0.4.5 Index]